Update license: LGPL => MPL 2.0
The LGPL has a virality problem with static linking that was not our
intention with the license choice for the standard library. The MPL 2.0
is much the same regarding derivative works but permits statically
linking with the standard library. Since we have no intention of
supporting dynamic linking, this is an important distinction!