all: use str for path components
On Linux, and other operating systems, path components can be any
sequence of bytes other than ASCII '/', and the stdlib API was
originally designed to accomodate this.

However, this is ultimately quite cumbersome. Some examples:

- argv can also be non-UTF-8, so os::args has to be updated as well,
  causing almost all Hare programs to deal with this dumb edge case
- Ditto for environment variables.
- Ditto for os::exec.
- Filenames can also contain \0, which breaks so much shit. Hare handles
  this as str but not as *const char.
- path::path cannot be passed to fmt for printing without lengthly
  special-case handling

This is a bit of unnecessary complexity that permeates huge swaths of
the stdlib and trickles into end-user programs. It simplifies things
greatly for everyone and for 99% of use-cases if we just abort when we
encounter non-UTF-8 paths.

That said, some programs do have a legitimate reason to support
non-UTF-8 paths, arguments, or environment variables. For this reason, I
have left the rt API intact, which understands both strings and byte
slices. Any program which has a genuine need to support this can write
syscalls or access rt's initialization vector directly. This is
non-portable - but that was a cost they were likely to bear regardless.