hare::module: refactor scan
This is mainly prep work. The "eligible" function has been removed,
because we need access to enough of its internals to no longer justify
it being separate. Future commits will continue this work by sorting the
list of inputs (to obtain a consistent module hash) and by filtering
them so that less-specific tagsets are discarded in favor of
more-specific sets.