unicode: initial module riggings
This defines constants for all of the Unicode properties found in the
UCD. I'll write a script later which converts the UCD XML representation
into a file which Hare can use to read these properties. This module
will also later be expanded with implementations of various algorithms
defined by Unicode.