hare::module: clean up candidate selection
This might be a bit weird. We may want to change the rules a bit on the
empty basename so that it never conflicts.

+linux/
+x86_64/

These should both be included. But, so should both of these:

+linux+x86_64/
+x86_64/

This only applies to the empty name, though. For the named file:

foo+linux.ha
foo+x86_64.ha

This should be incompatible, raising an error, but for this:

foo+linux+x86_64.ha
foo+x86_64.ha

The former should win.