os: add BUFSIZ; use for stdin, stdout, format::xml
I set this to 1MiB for now, for Linux. This is a lot bigger than musl (1
KiB) or glibc (8 KiB), but 1 MiB is not a log of memory to stick in .bss
these days, and, generally speaking, the larger the buffer is, the
better performance is. This change alone leads to a 30% increase in
format::xml performance, for instance.