bytes::index: implement two-way algorithm
This change implements a faster substring search algorithm for
bytes::index.  Instead of the naive solution we use the same main
algorithm that is also used in glibc and musl's strstr implementations.
There are some tweaks musl uses that this implementation lacks at the
moment, and we should probably implement another, simpler algorithm for
short needles.  Nevertheless, this is a significant step forward in
speeding up bytes::index.