hare::parse: handle pointer types in match cases
Earlier match_expr was assuming that * always means default case, which
is not always true for pointers.

Now we also check for => after *, and only then set as the default match
case. In other cases, parse this as a type and wrap it in a pointer
(because it's not possible to unlex 2+ tokens).

Minimal reproduction of the bug:

    match (x) {
    	nullable *foo => spam,
    	*bar => eggs,
    	* => void,
    };

Fixes #420