net::dns: refactor decoding for improved rdata
This abandons the stack-based API, which was only going to get more
complex and brittle (and hard to use) with the introduction of more
rdata formats, and instead just focuses on heap-allocating the decoded
message data. This also makes the rdata field use a more Hairy type
(tagged union of known rdata formats) and fully implements MX decoding.

Signed-off-by: Drew DeVault <sir@cmpwn.com>