encoding::base64: improve docs, trim down module
It might be a good idea to trim this further by removing some of the
functions which work with base64-encoded data as a byte slice. Since
base 64 is defined as a text format, it may make sense for this module
to only support representing it as a str (or indirectly, via an
io::stream).

Signed-off-by: Drew DeVault <sir@cmpwn.com>