cmd/hare: move archive files to the end of ld command
Fixes issues with linking when an archive is scheduled too early.

>From ld manual: "The linker will search an archive only once, at the
location where it is specified on the command line. If the archive
defines a symbol which was undefined in some object which appeared
before the archive on the command line, the linker will include the
appropriate file(s) from the archive."

To get symbols from archive to all other objects, archives must be
placed after everything else.

Signed-off-by: Alexey Yerin <yyp@disroot.org>