os::exec: use RDONLY on +freebsd
This is a bit of a doozy. It turns out that the earlier bug is that
O_EXEC interacts poorly with fdescfs. This uses O_RDONLY instead, which
is not perfect, but does "fix" the issue - assuming you mount fdescfs!

We will want to petition FreeBSD to mount fdescfs by default.