haredoc: lookup functions by @symbol
For example, in rt:

	@noreturn @symbol("rt.abort") fn _abort(msg: str) void;

Previously, running `haredoc rt::abort` would report a syntax error,
since `abort` is a reserved lexical token. Even if it wasn't, haredoc
still wouldn't yield any results. This commit allows alphanumeric
lexical tokens to be used in identifiers for looking up documentation,
and functions with an @symbol attribute will have that symbol checked as
well. This means that `haredoc rt::abort` now returns the documentation
for the _abort function.

Signed-off-by: Sebastian <sebastian@sebsite.pw>