encoding::base64: new design
base64 now uses an 'encoding' struct to represent the encoding alphabet,
which is more convenient to use than an array of characters.

This commit also introduces stream-based encoding.

Several of the previous design's functions accepted streams as
arguments. This was totally reworked, because if you're going to be
working with streams, you might as well create an encoder/decoder and
use io::read/write/copy. The convenience functions now only deal with
either strings or byte slices:

* encodeslice, decodeslice
* encodestr, decodestr

I've also updated the README to match the current design.

Signed-off-by: Ajay R <ar324@protonmail.com>
Signed-off-by: Drew DeVault <sir@cmpwn.com>