ascii: always check if strings are ASCII in strcmp funcs
The documentation for strcmp and strcasecmp states that if either string
isn't composed entirely of ASCII characters, then void is returned.
However, prior to this commit, if the lengths of the two strings didn't
match, this check wasn't performed on the remainder of the longer
string. The behavior of these functions now matches their documentation.

Signed-off-by: Sebastian <sebastian@sebsite.pw>