remove hash::finish
Previously there where two ways of obtaining a hash sum. Write,
sum, close; and write, finish. Sometimes this may lead to double freeing
the resources when defering the close, but using finish instead of sum.

This commit removes the finish method, so that there's only one way to
use the hash API to obtain the sum, and hence trying to avoid such
misuse of the API.

Signed-off-by: Armin Preiml <apreiml@strohwolke.at>