regex: improve performance
Previously, strings::sub was being called to get the content of each capture.
For scenarios like using findall() to find all occurrences of a substring in a
large file, this meant that the size of all runes in the string in bytes was
calculated for every submatch. Now, the size in bytes is calculated as we step
through the string, which means we know exactly from which byte to which we need
to slice to get a substring. This yields up to a 12x increase in performance
in my tests, probably more for very large files.

Signed-off-by: Vlad-Stefan Harbuz <vlad@vladh.net>