Handle negation in parse instead of lex
During the lexing stage, all number literals are taken to be unsigned.
The minus token is lexed on its own, independently of the number.
The sign of numbers is now determined during the parsing stage. This
fixes numerous edge cases, such as "1-1" incorrectly lexing as 1, -1.
This also simplifies the implementation of other language features, such
as parsing the index of a tuple access expression.

hare::parse doesn't check that the integer fits into the specified type.
That is, 1000u8 will successfully parse. This could be handled in
hare::parse with significant refactoring, but it's easier to handle it
during check in hare::unit.

hare::unit doesn't yet check for this, but hare::ast::number_constant
has been updated to aid with this when the time comes to implement it. A
sign field has been added, which will always match the sign of the
value, except when i64 overflows. This is necessary because
-9223372036854775808i64 is valid, but 9223372036854775808i64 isn't. In
the latter case, the value will overflow to a negative, but the sign
will remain positive, so hare::unit can detect the mismatch and report
an error.

Fixes: https://todo.sr.ht/~sircmpwn/hare/376
Signed-off-by: Sebastian <sebastian@sebsite.pw>