time::chrono: fix LOCAL init, improve docs
Renamed TZ_local to TZ_LOCAL. Consistent and less ungly, I suppose.

Improved docs for [[TZ_LOCAL]].

TZ_LOCAL.name now defaults to "Local". [[timezone]].name is meant to be
an identifier, rather than descriptive or end-user-friendly, and there's
a convention or unwritten rule to keep them whitespace-less for this
purpose. This convention is (will be) useful when parsing %L. Perhaps we
should add a .description field.

The filepath of the TZif file, if appropriate, is now used as part of
the name for [[TZ_LOCAL]].

Signed-off-by: Byron Torres <b@torresjrjr.com>