time::chrono,datetime: embed instant into moment
This commit facilitates a system of unification of all temporal types.
The time::chrono::moment type now:

* Embeds the time::instant struct type.
* Uses voidable .date, .time, and .zone fields.

Datetimes and moments are now intrinsically instants and inherit the
same properties of precision and ubiquity. Pointers to such objects may
be passed around seemlessly.

	*(&dt: *time::chrono::moment)
	*(&dt: *time::instant)

The .date and .time fields now hold observed values instead of
normalized values, and are to be evalutated and accessed with the new
exported functions listed below. The dangerous transform() function is
no longer needed.

New exports:

* fn time::chrono::getzone()
* fn time::chrono::getdate()
* fn time::chrono::gettime()
* fn time::chrono::from_datetime()

Removed exports:

* fn time::chrono::from_instant() // use chrono::new()
* fn time::chrono::to_instant()   // use *(&m: *time::instant)
* fn datetime::to_instant()       // use *(&dt: *time::instant)
* fn datetime::to_moment()        // use *(&dt: *time::chrono::moment)
* fn datetime::lookupzone()       // use chrono::getzone()
* fn time::chrono::transform()    // time::chrono now handles this
* fn datetime::transform()        // time::chrono now handles this