scripts/gen-docs.sh: use $(BINOUT)/haredoc
This way, HTML docs can be generated without running `make install`.

This makes `make docs/html` the correct way to generate docs, and
calling the shell script itself is now incorrect. Because of this, the
file has been renamed from gen-docs to gen-docs.sh, and its mode changed
from 755 to 644.

Signed-off-by: Sebastian <sebastian@sebsite.pw>