add crypto::rsa core functions
The core functions [[pubexp]] and [[privexp]] are ported from BearSSL.

I've chosen 4096-bit as maximum key size per default similar to BearSSL.
4096-bit is a reasonable size. Bigger keys will not improve security
that much in relation to how much more computational ressources they
require. If bigger keys are required, the [[BITSIZE]] parameter can be
changed at compile time.

This module requires the keys to be stored in byte slices, sinces they
are way simpler to handle. It allows to have a small number of init
functions, that prepare the keys and already do some basic checks.
On the other hand if keys are represented as structs, many more helpers
would be required to calculate the parameters, since not all key formats
store all parameters that we require. Key validation will need to go to
the core validations or into extra functions. That way the api would
have more moving parts. So the overhead of copying the parameters into byte
slices is worth trading for the simpler api design.

Signed-off-by: Armin Preiml <apreiml@strohwolke.at>