os: make stderr a handle instead of a file
Previously, stdin and stdout were handles, since they were really bufio
streams. Since stderr is unbuffered, it was just an io::file. This
commit makes stderr an io::handle, and adds stderr_file as an io::file.
stderr is still unbuffered; the handle is really the same as the file.
Besides just for consistency with the other streams, this change allows
all standard handles to be mutated into a different stream at runtime.
This is done by the new test runner to capture the output of tests, by
temporarily changing the value of os::stdout and os::stderr to a
bufio::dynamic stream.

Signed-off-by: Sebastian <sebastian@sebsite.pw>