special case libc implementation-dependent test
For input (NaN, Inf), math::complex::tanc128 returns
(0f64 with sign of NaN, 1f64 with sign of Inf). However, the sign of bit of
NaN is set under glibc, but it is not set under musl, resulting in test
failure.

This patch compares the test result using the comparison operator.
0f64 == -0f64 is true on both glibc and musl.

References:
https://lists.sr.ht/~sircmpwn/hare-dev/patches/33328
https://lists.sr.ht/~sircmpwn/hare-dev/%3C2I5P65667M8XC.2AAFMEP64ZWLL%40mforney.org%3E

Signed-off-by: Carlos Une <une@fastmail.fm>