bufio::scan_rune: properly read last 3 bytes
bufio::scan_rune returns io::EOF if it encounters an io::EOF during the
scan_readahead call. However, this call is made when the pending bytes
are < 4. The result is that if you will read EOF when you have 3 bytes
left in the buffer, leaving them unaccessible.

Only return io::EOF if the scan.pending is 0 after the call to
scan_readahead. Add a test for this behavior.

Signed-off-by: Tim Culverhouse <tim@timculverhouse.com>