Makefile: Pass LDLINKFLAGS to link hare
Since hare(1) invokes $(LD) with $(LDLINKFLAGS) but hare(1) itself is
built with a direct $(LD) invocation from the makefile, it was linked
without those (usually disto-wide) flags. The flags are now added to
the $(LD) command line to remedy this situation.

The same is done for hare-test(1) for consistency, even though this
program is not installed.

Signed-off-by: Dridi Boukelmoune <dridi.boukelmoune@gmail.com>