error: Align located messages with other toolchains
The make(1) integration in Vim can identify error messages with a
location and build a list of them that can then be navigated. When
make(1) is executed directly from vim, for example with `:make` or
`:make check` such error messages will be registered and the editor
will switch the current buffer to the first error location.

This integration requires error messages to be in the form of:

    file:line:column: description

Vim switches to the `file` file and moves the cursors to line `line`
and column `column`, showing `description` in its status bar, which
is very convenient.

The test case was updated for consistency, to represent the source
code position in the same file:line:column form. Test failures for
the lexer don't result in a "located" error message anyway.

Signed-off-by: Dridi Boukelmoune <dridi.boukelmoune@gmail.com>