time::chrono: use daydate, daytime; improve docs
This commit renames the .date and .time fields of
[[time::chrono::mooment]] to .daydate and .daytime, to disambiguate
between the "date" noun's reserved meaning of a general date, and a
scalar day number date; and between the problematic multiple definitions
of the "time" noun and the time-of-day. "time" is too general to be
anything but a module name.

Various docs are updated, and the READMEs are made terser.