Return tuple directly from strings,bytes::cut,rcut
strings::cut et al are convenience functions which aim to address the
common 95% of cases, an approach which is common to much of the standard
library's design. It is not important for this interface to be
exhaustive; other tools are available for those who need to treat the
presence or absence of the delimiter differently. The convenience of
this convenience function is greatly diminished should the 95% of users
who do not need to distinguish these cases be required to add `as (str,
str)` -- a full 25% of the 80-character line width budget -- for every
call.

This reverts commit da442e0bf76cac19a137a3f779b5e0d838b94c8a.
This reverts commit aa9d6b57fed162be8d5d1c59ef3fb0614e504bba.