os::exec: only open() regular files
Currently, os::exec::open() will happily return success when e.g.
opening a directory (as long as the user has execute permission on it).
This can become a problem, as os::exec::lookup() uses a successful call
to open() as indicator that it found the right executable.

In the situation where a directory that is also early in PATH contains
e.g. another directory with the same name as some executable, that
directory will be returned instead of the executable. Making open()
check its target more diligently - and hence fail early - solves that.

Simple reproduction example:

    mkdir -p test/hare
    export PATH="$(pwd)/test/:$PATH"
    haredoc -h

Signed-off-by: Conrad Hoffmann <ch@bitfehler.net>