strconv: improve error messages
When combined with the previous patch, this makes it so when a negative
number is used where an unsigned integer is expected, the error message
doesn't say "Input string is not an integer"; it correctly states that
the requested number doesn't fit.

Likewise, when invalid is returned from stof64 et al, the error message
doesn't say that "Input string is not an integer".

Signed-off-by: Sebastian <sebastian@sebsite.pw>