hare::parse: improve struct/union/enum comments
Comments that appear immediately after a struct/union member now
associate with said member, rather than the one after it. Enum comment
behavior was also slightly tweaked to match that of structs and unions
(if both a comment above and a commment on the same line are present,
the comment above is used).

Signed-off-by: Sebastian <sebastian@sebsite.pw>