cmd/hare: Bring back LDFLAGS support
The hare(1) command used to pass additional flags to the linker using
the LDFLAGS variable but it conflicted with how the LDFLAGS variable is
traditionally used, linking with the cc(1) command.

Since hare(1) may use either ld(1) or cc(1) for linking, LDLINKFLAGS can
not work as a one-size-fits-all.

By adding a formal cc_tool we can support both LDFLAGS and LDLINKFLAGS.