path: exclude nul terminator from PATH_MAX
On Linux, PATH_MAX is 4096 in C, but this includes the terminating NUL
byte. Hare strings don't have a terminating NUL byte, so it's possible
for a path whose length is PATH_MAX to actually be too long. This is
corrected here by changing PATH_MAX to 4095.

The only place the terminating NUL byte is relevant is syscalls, but rt
provides its own PATH_MAX which is 4096 bytes anyways.

Signed-off-by: Sebastian <sebastian@sebsite.pw>