unix::signal+freebsd: s/POLL/IO/g
So, SIGPOLL is specified by POSIX, but FreeBSD doesn't support it. Ok,
fine, whatever. But some things in FreeBSD are documented as though
SIGPOLL is supported. Like, the code enum this commit touches, this was
copied directly from FreeBSD's sys/signal.h. And the comments there just
straight up reference a signal which doesn't exist.

On Linux, SIGPOLL is synonymous with SIGIO, so I updated the docs here
to reference that instead.

Signed-off-by: Sebastian <sebastian@sebsite.pw>