Add TREES
This document is used to keep track of various non-upstream trees around
the ecosystem that may be of interest to a contributor or group of
contributors.

It follows the format of MAINTAINERS, which is updated to add a couple
of field types to better accomodate this use-case. The "B" entries (for
bug trackers) is copied from the Linux maintainers, and "N" (free-form
notes) is origanal to Hare.