cmd/hare: Rename the temporary file after writing typedefs
Running multiple build drivers in parallel should no longer cause
"expected non-void prefix in get_cache()" assertion failures.

While renaming a file might not seem like it would release a lock, that's what
it does. The build driver does not hold a reference to a file when re-trying to
lock it and just uses a path. When a file is moved, there would be nothing at
the original path, so the file is created and will be treated as unlocked.

That's not a problem for most cache types, but stage::SSA also writes a
typedef pointer to <out>.td, yet it does so *after* the lock is released by
renaming. This violates the expectation that the cache is ready after lock is
released - there are no typedefs yet. get_td is happy and just ignores the
error, but then a hash is not written into ctx.hashes, which later causes an
assertion failure in get_cache.

Signed-off-by: Alexey Yerin <yyp@disroot.org>