fs, os: improve some docs
The main motivation for this is discouraging the use of flag::CREATE
with open, instead directing the user to create. It also clarifies in
create's docs that the file is only created if it doesn't already exist.

This also corrects create's docs, by documenting that flag::TRUNC is
used by default. fs::flag didn't list TRUNC as one of the mandatory
flags for an fs::fs to support though, so I had to add that.

Signed-off-by: Sebastian <sebastian@sebsite.pw>