hare::unparse: overhaul with synfunc
Most functions in hare::unparse now take in a 'syn' parameter, which is
a pointer to a function that's called for each thing unparsed. This
gives the user lots of freedom for how the AST should be unparsed.

Two synfuncs are provided in hare::unparse: syn_nowrap and syn_wrap,
which don't do any styling, but the latter wraps long lines.

The bare minimum was done to get haredoc compiling; haredoc is made to
fully take advantage of this in a future commit.

Closes: https://todo.sr.ht/~sircmpwn/hare/842
Signed-off-by: Sebastian <sebastian@sebsite.pw>