encoding::pem: restructure decoder
Prior to this commit, pemdecoder used an awful hack in which, when
initialized, the b64 field was forcefully zeroed, and a bool b64_ready
was used so b64 could actually be set the first time the stream was read
from. This was necessary since a base64::decoder needs to be initialized
with an *io::stream, but the io::stream being used was part of the
pemdecoder which hasn't been returned yet, so its address couldn't be
taken.

This is solved by storing a bufio::stream (non-pointer) in b64stream,
and storing and initializing a b64stream in decoder (replacing the
bufio::stream in decoder). This way, since the stream is stored in the
decoder which is already owned, the address can be taken for the
base64::decoder in the pemdecoder. This also allows the same b64stream
to be reused, rather than creating a new one every time `next` is
called.

Additionally, the 'in' field of pemdecoder can be removed, since it just
duplicates a field already stored in base64::decoder.

Signed-off-by: Sebastian <sebastian@sebsite.pw>