Move string comparison logic to sort::cmp::strs()
This way, strings::compare() becomes a wrapper for sort::cmp::strs()
instead of the other way around. That in turn makes strings depend on
sort::cmp instead of the other way around, making sort::cmp (and by
extension, sort itself) usable by bytes (and potentially other modules
that strings depends on).

Thanks to Seb for the idea for this approach, see
https://lists.sr.ht/~sircmpwn/hare-dev/%3Cd2d241c3-e405-49bf-b874-660f7cfa50fd%40bitfehler.net%3E

Signed-off-by: Conrad Hoffmann <ch@bitfehler.net>