hare::lex: take in scanner
The lexer previously was initialized with an io::handle, and then
created and managed its own bufio::scanner internally.

This meant that there was no reliable way to know where a stream will be
positioned after lexing is completed. So it's not possible to read from
a handle after lexing from it.

This is solved by changing the interface to receive a scanner which is
managed by the user. This way, code can use a scanner, lex something
from that scanner, and then continue to use the same scanner, with the
guarantee that the position will be immediately after the last token
lexed.

This is useful in e.g. hare::parse::doc, which scans a comment, and when
"[[" is encountered, it parses an identifier from the scanner, and then
continues scanning. If the lexer interface weren't changed, this code
would instead need to allocate a buffer to store the identifier, then
pass a memio::fixed stream to the lexer, which is wasteful since the
lexer is internally using a scanner anyway, so there's no reason it
shouldn't be able to share the same scanner.

In addition, the internal storage of the lexer is overhauled in this
commit, to take advantage of the scanner for some information, and to
minimize the size of fields elsewhere. The effect in both cases is less
information needs to be stored in the lexer itself, which also
simplifies the implementation in various places.

Signed-off-by: Sebastian <sebastian@sebsite.pw>