bufio/scanner: fix slice or array access out of bounds
I struggled with the scanner used with pipes in the context of himitsud
and hiprompt-tty.

The problem come from the scan_bytes methods. It suppose z is the
readahead size, but in fact this is the previous scan.pending. This
means that multiple loop of this scope cause nread to become higher than
scan.pending, and so cause a bound check abort.

To fix this, we just use z as nread, renamed as prevpending to avoid mistakes.

Signed-off-by: Willow Barraco <contact@willowbarraco.fr>