strings: consistently use fromutf8_unsafe
strings used to use the unsafe variant in some places, and the safe
variant in other places, with no pattern as to which is used where. I've
changed all fromutf8 calls to fromutf8_unsafe calls, so calls to
strings:: functions don't repeatedly iterate through the string to check
that it's valid UTF-8 when it's already been proven that it is.

Signed-off-by: Sebastian <sebastian@sebsite.pw>