time: change signature of *_to_timespec funcs
Also fixes a memory management "bug" in sleep. The variable req was set
to the address of a variable local to the loop, which works as of now
but which I assume is undefined behavior (the spec is currently silent
about this so it's tough to say).

This is a breaking change: existing code can be migrated by taking the
value of the function result, rather than creating a binding and passing
it in as a pointer.

Signed-off-by: Sebastian <sebastian@sebsite.pw>