OpenBSD: fix net/ calls to rt functions
during the port to OpenBSD i converted a lot of rt:: networking
functions to use void as a return type instead of int just to make
things cleaner. i forgot to change some of the wrappers in net::,
this fixes that. other platforms should also be changed, just haven't
found time to do that.

Signed-off-by: Lorenz (xha) <me@xha.li>