fmt: Fix fmt::mods documentation
With the changes to padding made in d09ea3b72 (2023-11-06), this example
became inaccurate as mods.pad defaults to NUL. This is difficult to
notice on my terminal at least, the NUL-byte is invisible.

Once (if?) @default functionality is added, it might be a good idea to
set the pad field to have a default value of ' '.

Signed-off-by: Sebastian LaVine <mail@smlavine.com>