rt: BPF socket structs & option levels for Linux
I've been working on a BPF module with my own versions of these structs
and definitions, but I figured it was worth issuing a patch to get
them included in the Linux runtime here. I kept the names of the struct
fields as close as possible to the kernel versions.

Signed-off-by: Blain Smith <rebelgeek@blainsmith.com>