asn1: oid support
encoding/asn1/oiddb/db.txt contains all oids required by the stdlib. It
is in a separate module so that projects may create their own oid list
by appending to the stdlib and generating the db with the genoiddb
command.

It is still no ideal solution if multiple libs will require their own
oids added to the db.

Signed-off-by: Armin Preiml <apreiml@strohwolke.at>