asn1: add decode support for non utf8 string types
numeric and ia5, which is basically ascii, are required by some
entries in x.509. UniversalString, BMPString and T61 are only provided
for legacy support.

Note that this only supports a subset of T61. There are still some
certificates in the mozilla trust store that have T61 encoded strings.
Though the characters may only be a subset of ASCII.

Signed-off-by: Armin Preiml <apreiml@strohwolke.at>