haredoc: warn and continue on parsing error
This only applies to -Ftty; -Fhtml is unchanged.

Prior to this commit, haredoc would exit with a failure status if any
file failed to parse. Now, when a file fails to parse, haredoc will
display a warning showing all the parsing errors that occurred, but it
won't terminate; it will still show documentation from all files that
parsed successfully.

Signed-off-by: Sebastian <sebastian@sebsite.pw>