NetBSD: make the build pass, replace all compat syscalls
compat_13 is a non-builtin kernel module that can be autoloaded if a
compat_13 syscall is called, but the file belongs to the `mi` set and
isn't available in all installations.

The other compat modules are builtin to the GENERIC kernel so they
should be available, but we may as well replace them.

- Replace SYS_compat_13_sigaction13 with SYS___sigaction_sigtramp
- Replace SYS_compat_13_sigprocmask13 with SYS___sigprocmask14
- Replace SYS_compat_43_fstat43 with SYS___fstat50
- Replace SYS_compat_30_getdents with SYS___getdents30
- Replace SYS_compat_50_wait4 with SYS___wait450
- Replace SYS_compat_50_nanosleep with SYS___nanosleep50
- .builds/netbsd.yml: Update the path to the leapseconds file
- .builds/netbsd.yml: Use sudo not doas