unix::passwd: use bufio::scanner and refine API
Programs which enumerate all users or groups on the system are affected
by the breaking change. The API is refined such that the user must
create a parser object, from which each entry in the passwd or group
file is borrowed while iterating. Users of one-off functions (e.g.
unix::passwd::getuid) are not affected.

This is a breaking change.

Signed-off-by: Drew DeVault <sir@cmpwn.com>