net/uri: fix silly usage of wantrune()
This is a follow-up to my previous fix. I now realized that using
`wantrune()` like I did here is silly. If I understand correctly,
`wantrune()` expresses that there should be a next rune or the URI is
invalid, but this is not the case here. So stick to `strings::next()`
and handle `done` instead of `invalid` (not much sense in handling an
event called `invalid` and continue parsing...).

Signed-off-by: Conrad Hoffmann <ch@bitfehler.net>