bufio::scanner: add options to fine-tune the behavior of a scanner
EOF_GREEDY is the previous scanner behavior. Now with the new default
value, the scanner will return EOF if it reachs the source EOF without
encountering the token.

This new option is necessary to parse uncomplete buffers in a
non-blocking way (ex: in hare-ev). It is also necessary to parse
ill-formed data in general.

This is a breaking change. Programs that relies on the previous behavior
may now have a confusing scanner result that might be difficult to
debug.

The changes to the tests represent correctly the change of behavior:

unix::hosts::next need line breaks, because /etc/hosts-formatted formated lines
ends with line breaks. "127" is not invalid, it is an incomplete line.
The same conclusion for format::ini.

Signed-off-by: Willow Barraco <contact@willowbarraco.fr>