Makefile: Build haredoc with $(HAREFLAGS)
I didn't notice this problem when I introduced the HARE_BUILD_ENV
variable to deduplicate the `hare build` logic, but it existed prior
to my change. Meanwhile harec2 went away and there wasn't much left
to deduplicate. HAREFLAGS is not part of the environment variables
recognized by the hare(1) command and is rather something that should
be accounted for in the build system. This was the case, but it was
done inadequately.

>From now on $(HAREFLAGS) should be passed as command line arguments to
be honored. The $(HARE_BUILD_ENV) variable can remain, it will be used
in the next commit. On the other hand, $(HARE_DEFINES) are not needed
by haredoc(1).

Refs 605a70a66aa6c34a0aee9929a54cfc9ed95575e8

Signed-off-by: Dridi Boukelmoune <dridi.boukelmoune@gmail.com>