Makefile: Install hare(1) built by hare
The hare(1) command is bootstrapped from direct calls to harec(1) and
the rest of the toolchain so as a result it does not honor HAREFLAGS
among other things done by a regular hare build. This prevents software
distributions from shipping proper release builds using the -R option.

To avoid this pitfall, a new hare-install program is built using hare at
install time, and is installed as $(BINDIR)/hare. The benefits include a
consistent build of installed hare(1) and haredoc(1) commands and an
implicit check that hare(1) successfully builds itself.

Signed-off-by: Dridi Boukelmoune <dridi.boukelmoune@gmail.com>