cmd/hare: Print dependencies to a t
The new 'hare deps -t' option changes the output to a plain text one.
This format can easily be consumed by field-based programs such as
tsort(1) (to a t!) or the more frequent utilities like sh(1) (while
read loops) or awk(1).

To illustrate how it renders, here is the output for the hare-compress
package, as of a4bb204305ef67a39dc1af3e8e843a2a4f5bee45:

    $ hare deps -Dst ~/src/hare-compress/
    compress::flate endian
    compress::flate errors
    compress::flate io
    compress::gzip bufio
    compress::gzip bytes
    compress::gzip compress::flate
    compress::gzip endian
    compress::gzip errors
    compress::gzip hash
    compress::gzip hash::crc32
    compress::gzip io
    compress::gzip memio
    compress::gzip time
    compress::zlib compress::flate
    compress::zlib endian
    compress::zlib errors
    compress::zlib hash
    compress::zlib hash::adler32
    compress::zlib io

The first column gives the effective list of modules provided by this
package. The second column gives required modules, once we filter out
the provided modules.

There is no compress module, because there are no sources. Adding such a
source file without depending on other modules would slightly change the
output:

    $ cat >~/src/hare-compress/compress/version.ha <<EOF
    export const version_string = "hare-compress x.y.z";
    EOF
    $ hare deps -Dst ~/src/hare-compress/
    compress -
    compress::flate endian
    compress::flate errors
    compress::flate io
    compress::gzip bufio
    compress::gzip bytes
    compress::gzip compress::flate
    compress::gzip endian
    compress::gzip errors
    compress::gzip hash
    compress::gzip hash::crc32
    compress::gzip io
    compress::gzip memio
    compress::gzip time
    compress::zlib compress::flate
    compress::zlib endian
    compress::zlib errors
    compress::zlib hash
    compress::zlib hash::adler32
    compress::zlib io

Note the presence of a 'compress -' line expressing the presence of
the module and its lack of dependencies.

The motivation is to enable module processing in package managers,
in the non-recursive mode, to derive virtual provides from the first
field and virtual requires from the second one, from the authoritative
tool chain.

Signed-off-by: Dridi Boukelmoune <dridi.boukelmoune@gmail.com>