time::chrono: fix loading TZ filepaths
The doc says:

> If [name] is a full filepath (begins with '/'), it is used directly instead.

But the function was still prefixing with TZDB_PATH when using an
absolute path.

This cause a huge problem on Alpine linux, when the user uses the
setup-timezone without the -i. The setup would copy the timezone file
from /usr/share/zoneinfo/** to /etc/zoneinfo/**. Meaning date::localnow
would just fallback to UTC.

Signed-off-by: Willow Barraco <contact@willowbarraco.fr>
Signed-off-by: Byron Torres <b@torresjrjr.com>
Co-authored-by: Byron Torres <b@torresjrjr.com>