math: remove generic float functions
f32 can promote to f64, and the generic functions always returned f64
(rather than f32) anyways. The only time the generic functions are
useful is when you have a types::floating value which you don't know the
actual type of. In that case, you now have to match on it yourself. I
think this is a worthwhile tradeoff for a cleaner API.

This is a breaking change.

Signed-off-by: Sebastian <sebastian@sebsite.pw>