math: remove generic int functions
Similar to the generic float functions, all supported types can promote
to i64, so the only time the generic functions are useful is when have a
already have a types::integer value which you don't know the type of.
This was the case in fmt::, which now has to match on the value itself.
fmt:: is an exception to the norm though; I doubt most users will have
any issues migrating their code to use the appropriately typed functions
instead.

This is a breaking change.

Signed-off-by: Sebastian <sebastian@sebsite.pw>