hare::doc: don't introduce own error type
The problem with this is that Harec used to (incorrectly) do a full
dealias upon encountering a ... operator, and doc::error relied on this
behavior in function return types. This could be fixed by introducing
some casts, but there is really no good reason to have another separate
error type there at all.

Signed-off-by: Bor Grošelj Simić <bgs@turminal.net>