regex: fix `jump_idxs` interference
Take this case: a|(b)

We need to remember the position of the `|`, because we will need it
when we reach the end of the string.

However, when we encounter the `)`, all `jump_idxs` are deleted; uh oh!

This commit makes it so that `jump_idxs` are stored per `group_level`,
so as not to interfere with each other.

Signed-off-by: Vlad-Stefan Harbuz <vlad@vlad.website>